Разолжил классы по соотв. папкам и файлам независимо. Внес правки для перегрузки операторов
