+++ src/spacebattle/commands/__init__.py
@@ -1,6 +1,8 @@
 from .command import Command
-from .default_command import DefaultCommand
-from .put_command import PutCommand
-from .retry2_command import Retry2Command
-from .retry_command import RetryCommand
-from .write_command import WriteCommand
+from .default import DefaultCommand
+from .move import MoveCommand
+from .put import PutCommand
+from .retry import RetryCommand
+from .retry2 import Retry2Command
+from .rotate import RotateCommand
+from .write import WriteCommand
