+++ src/spacebattle/commands/default.py
