+++ src/spacebattle/common/__init__.py
@@ -1,0 +1,2 @@
+from .angle import Angle
+from .vector import Vector
