+++ src/spacebattle/common/angle.py
@@ -7,26 +7,17 @@ class Angle:
             assert type(d) is int
             assert type(n) is int
             assert d < n, "Угол поворота должен 'd' быть меньше кол-ва частей 'n'"
-        self.__d = d
-        self.__n = n
-
-    @property
-    def d(self) -> int:
-        return self.__d
-
-    @property
-    def n(self) -> int:
-        return self.__n
+        self.d = d
+        self.n = n
 
     def __add__(self, other: "Angle") -> "Angle":
         """Перегрузка оператора '+' для сложения углов"""
-        assert self.__n == other.n, "Кол-во частей 'n' должно быть одинаковым"
-        self.__d = (self.d + other.d) % self.__n
-        return self
+        assert self.n == other.n, "Кол-во частей 'n' должно быть одинаковым"
+        return Angle(d=(self.d + other.d) % self.n, n=self.n)
 
     def __eq__(self, other: "Angle") -> bool:
         """Перегрузка оператора == для сравнения углов"""
-        return other.d == self.__d and other.n == self.__n if type(other) is Angle else False
+        return other.d == self.d and other.n == self.n if type(other) is Angle else False
 
     def __str__(self) -> str:
-        return f"({self.__d}, {self.__n})"
+        return f"({self.d}, {self.n})"
