+++ src/spacebattle/objects/__init__.py
@@ -1,0 +1,2 @@
+from .moving import MovingObject
+from .rotating import RotatingObject
