+++ src/spacebattle/objects/moving.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from src.spacebattle.vector import Vector
+from src.spacebattle.common.vector import Vector
 
 
 class MovingObject(ABC):
@@ -17,13 +17,3 @@ class MovingObject(ABC):
     @abstractmethod
     def get_velocity(self) -> Vector:
         ...
-
-
-class MoveCommand:
-    """Класс реализующий движение"""
-
-    def __init__(self, obj: MovingObject) -> None:
-        self.__obj = obj
-
-    def execute(self):
-        self.__obj.set_location(self.__obj.get_location() + self.__obj.get_velocity())
